[NO-TICKET] Upgrade "datadog-protos"/"ddketch-agent" to unlock tonic 0.12.3
**What does this PR do?**

This PR upgrades the libdatadog dependencies coming in from the
"saluki-backport" repository: datadog-protos and ddsketch-agent.

These unlock the upgrade of tonic to 0.12.3 which will solve
the following security advisory:
https://github.com/DataDog/libdatadog/security/dependabot/22 .

**Motivation:**

Get rid of outdated dependencies.

**Additional Notes:**

The upgrade of stuff from "saluki-backport" brings other changes;
I'll sync with @duncanpharvey to make sure this is good to go.

**How to test the change?**

Not sure -- hopefully existing test coverage is enough?
I'll sync on this with @duncanpharvey as well.