Implement UnderlyingBytes for ByteSlice so it can be borrowed as underlying for TinyBytes::Bytes (#764)
Update data-pipeline's send function to accept a tinybytes::Bytes for the payload type. 

Also update the data-pipeline-ffi ddog_trace_exporter_send function to transmute the incoming ByteSlice to static. The ByteSlice is expected to live for the length of the function call, but it is technically possible to borrow the underlying data for longer, leading to stale references. This will be addressed as part of APMSP-1621.