Change profiling build to first generate the new header in ./target, and then run deduplication
Otherwise we were having issues when rerunning built after deleting the destination directory because of build.rs script caching.

Also somehow the data-pipeline–ffi symbols were included in the build by default, but the headers were only generated when the ARG_FEATURES flag was passed.
So I removed data-pipeline-ffi from the defaults because without the headers, it doesn't make much sense to include it in the built shared library