Change profiling build to first generate the new header in ./target,and then run deduplication (#798)
# What does this PR do?

* Change profiling build to first generate the new header in ./target,and then run deduplication
* Remove the `data-pipeline-ffi` feature from the default features used in the shell build script

# Motivation

The way we currently do header generation is not great.
The build.rs directly emit the headers in the final output directory of the script. They are then modified in place to deduplicate headers.

This has multiple issues on unclean builds:
* If we delete the final output dir, but haven't modified the rust code, the headers are not re-generated and the build fails
* If we run the script with multiple features, we will leave previous header files there which creates error

The changes makes it so that:
* The bindgen build.rs scripts always output to ./target/inlcude
* We then copy the headers that we care about for the current set of feature flags to the final output directory
* We run header deduplication on the copied header

Also somehow the data-pipeline–ffi symbols were built by default in the profiling lib, but the headers for the symbols were only generated when the ARG_FEATURES flag was passed. Now the data-pipeline headers are always generated