Allow intern to fail cleanly rather than blowing up
In particular, in the unlikely event that we would overflow the id,
we signal an error back to the caller rather than impacting the
application.

The caller is expected to stop using this string table and create
a new one instead. In the future we may make it easy to do so, by
e.g. having an API to create a new table from the existing strings
or something like that.