Make `intern_new` private
Adding it as `pub` was an oversight, since the intent is for this to
be an inner helper that's only used by `intern` and by `new`.

Having this as `pub` is quite dangerous as this method can easily
be used to break a lot of the assumptions for the string storage.