Remove redundant conversion of `ManagedStringId`s to... themselves
Not sure if this was there from the beginning or the result of
successive refactors, but yeah, we were unpacking and repacking
the `ManagedStringId`s uselessly (the input and output types were
the same!).

I'm pretty sure the compiler would optimize all of this away -- in
the end this is a struct with an int -- but our code sure does look
better.