Fix possible deadlock with remote configuration expiring during insertion (#888)
* Fix possible deadlock with remote configuration expiring during insertion

In this case the dead handler (internally holding a lock to the service map) attempting to remove (needs to acquire the configurations map) while add_runtime (which holds a lock to the configurations map) is called (needs to acquire the services map).

Also doing a trivial re-order in session_info to not retain remote config invariants across runtimes and application locks.

Signed-off-by: Bob Weinand <bob.weinand@datadoghq.com>

* Print memory statistics on memory exceeded

Signed-off-by: Bob Weinand <bob.weinand@datadoghq.com>

* Fix possible deadlock with simultaneous remote config insertion and removal

The services lock must never be acquired before the runtimes lock.

Signed-off-by: Bob Weinand <bob.weinand@datadoghq.com>