[PROF-11405] Fix managed string storage causing corrupt profiles (#896)
**What does this PR do?**

This PR fixes a bug with managed string storage's cache invalidation
logic that caused corrupt profiles.

Specifically, profiles could be emitted that referenced wrong or even
non-existing entries in the string table, which would fail to parse.

**Motivation:**

Fix managed string storage bug.

**Additional Notes:**

Prior to this PR, the managed string table was using the profile's
`StringTable` pointer as a key for cache invalidation in the
`ManagedStringId` to `StringId` mapping.

This is not correct since when a profile gets reset, the `StringTable`
pointer **doesn't change**.

This bug lay dormant because Ruby uses two profiles and switches
back and forth between them. This meant that in most cases the
cache would go back and forth between the two profiles, working
correctly... until a cache entry was left behind and got reused
incorrectly.

**How to test the change?**

I've added a test to make sure we don't regress here.