Make `get_seq_num` require `mut self`
As far as I can tell, the previous implementation was hiding a data
race. Specifically: if two threads concurrently called `get_set_num` for
the same id, they would fight when trying to mutate the cache, which
very much looks to me might result in an invalid result.

To avoid this, let's make sure to mark this method as mutable.

Anyway I don't believe this will impact performance as right now
we don't particularly have concurrent **read** accesses to the managed
string table that would take advantage of sharing the read lock.