Replace `RwLock` with `Mutex` for managed string storage
Almost all operations now require a write lock, so for now let's move
to using a regular `Mutex` as we're not particularly getting anything
still from the `RwLock`.

(As we crystallized our use of this feature and its API, more and more
things that appeared like they could take advantage of read locks
actually needed a write lock, so that's how we got here).