[PROF-11405] Small managed string table fixes + cleanups (#905)
* Mark `unintern` as requiring a mutable reference

We already use write locks to access this, but mistakenly forgot to
mark this method as requiring a mutable reference. (So this was a
loaded gun that we hadn't fired...)

* Use checked operations when touching usage_count

* Minor: Simplify error handling by using `context`

* Minor: Refuse to read string back if refcount is zero

* Make `get_seq_num` require `mut self`

As far as I can tell, the previous implementation was hiding a data
race. Specifically: if two threads concurrently called `get_set_num` for
the same id, they would fight when trying to mutate the cache, which
very much looks to me might result in an invalid result.

To avoid this, let's make sure to mark this method as mutable.

Anyway I don't believe this will impact performance as right now
we don't particularly have concurrent **read** accesses to the managed
string table that would take advantage of sharing the read lock.

* Replace `RwLock` with `Mutex` for managed string storage

Almost all operations now require a write lock, so for now let's move
to using a regular `Mutex` as we're not particularly getting anything
still from the `RwLock`.

(As we crystallized our use of this feature and its API, more and more
things that appeared like they could take advantage of read locks
actually needed a write lock, so that's how we got here).