[PROF-11405] Replace managed string storage Rc with Arc (#919)
* [PROF-11405] Replace managed string storage Rc with Arc

**What does this PR do?**

This PR replaces the
[`Rc`](https://doc.rust-lang.org/std/rc/struct.Rc.html) being used to
wrap the managed string storage with an
[`Arc`](https://doc.rust-lang.org/std/sync/struct.Arc.html).

As @danielsn pointed out, having `Rc<Mutex>` is suspicious, and he
was right: this `Rc` can and will be accessed concurrently by
users of libdatadog and thus it needs to be an `Arc`, as otherwise
the `Rc`'s internal counter is subject to racy reads an writes.

**Motivation:**

Make managed string storage production-ready.

**Additional Notes:**

We probably got "lucky" with this not causing visible issues so far
because there's not a lot of concurrency going on with this type.

**How to test the change?**

The existing tests validate the API continues to work with this change.

* Add safety comment