[PROF-11524] Bump libdatadog version to 17.0.0 in preparation for release
**What does this PR do?**

This PR bumps the libdatadog version to 17.0.0 in preparation for
release.

We're going from 16.x to 17.0.0 because of a number of small
backwards-incompatible changes, including
https://github.com/DataDog/libdatadog/pull/948 .

**Motivation:**

Release libdatadog 17.0.0.

**Additional Notes:**

I'm opening this PR before
https://github.com/DataDog/libdatadog/pull/913 and
https://github.com/DataDog/libdatadog/pull/948 are merged in
but I'll wait for both of them to hit `main` before merging this.

**How to test the change?**

I've tested the libdatadog 17 release locally using the Ruby profiler
test suite.