[NO-TICKET] Add runtime_platform tag automatically to reported profiles
**What does this PR do?**

This PR adds the `runtime_platform` tag as per
https://docs.google.com/spreadsheets/d/1LOGMf4c4Avbtn36uZ2SWvhIGKRPLM1BoWkUP4JYj7hA/edit?gid=0#gid=0
in the same format "architecture-os[-musl]".

**Motivation:**

Up until now, only Ruby was reporting this tag, but in incident 35390
it would've been useful to have this information, and it seems
easy enough to add.

**Additional Notes:**

N/A

**How to test the change?**

I've added test coverage for this feature. I've also tested it
with the Ruby profiler.