[NO-TICKET] Include libdatadog profiling crate version when reporting profiles (#957)
**What does this PR do?**

This PR tweaks the `internal` metadata section that gets reported along
with every profile to include the libdatadog profiling crate version
(which matches the libdatadog version currently).

**Motivation:**

In incident 35390 I had to manually check which libdatadog version
a given profiler version was reporting and I missed having a nice
unified way to figure this out.

**Additional Notes:**

Arguably this could go in the `info` version instead, but I decided
to keep it as internal info for now, and we can always change it later.

**How to test the change?**

I've updated the unit tests and also validated this works with
the Ruby profiler.