[PROF-11524] Fix not being able to set profile start_time at serialization
**What does this PR do?**

In https://github.com/DataDog/libdatadog/pull/926 we removed the
`start_time` argument from
* `ddog_prof_Profile_new`
* `ddog_prof_Profile_with_string_storage`
* `ddog_prof_Profile_reset`

The intention was that having it as an argument in
`ddog_prof_Profile_serialize` was enough, and anyway almost
everyone was passing in `null`s in the APIs above.

I missed when suggesting in that PR that the `start_time` for
serialize was the `start_time` for the next profile, not the one
being serialized.

In this PR I'm changing that behavior: the `start_time` argument
to serialize now controls the time for the profile being
serialized, allowing the profiling library to have exact control
over this value.

I've also removed the duration.

[As can be seen in this github
search](https://github.com/search?q=org%3ADataDog+ddog_prof_Profile_serialize&type=code)
this is not expected to impact anyone: everyone's passing
`NULL` for `start_time` and `duration` when calling serialize already.

**Motivation:**

Allow Ruby profiler to set the start_time of profiles.

**Additional Notes:**

N/A

**How to test the change?**

I've tested this with my experimental libdatadog 17 branch for Ruby.