feat(profiling): add runtime_platform tag automatically (#954)
This PR adds the `runtime_platform` tag as per [Profiler common tags](https://docs.google.com/spreadsheets/d/1LOGMf4c4Avbtn36uZ2SWvhIGKRPLM1BoWkUP4JYj7hA/edit?gid=0#gid=0). It uses the target triple, which is a bit different from what Ruby did before, but close enough.

# Motivation

Up until now, only Ruby was reporting this tag, but in incident 35390 it would've been useful to have this information, and it seems easy enough to add.

# Additional Notes

We should probably talk to the Datadog platform team to figure out if there's a "standard" or "approved" tag name, but this one is already in use so let's not block on that. We can migrate later as needed.

# How to test the change?

I've added test coverage for this feature. I've also tested it with the Ruby profiler.