build: update cc, ring, and remove spin (#971)
The ring crate has dependabot alert on it. It requires an updated
`cc` crate, which has been problematic because the 1.2.x series has
been incompatible with our *-alpine-linux-musl builds. They recently
merged and releases a fix (or at least, it should be fixed), so we
can now update cc too.

These changes mean spin is no longer used, so it's dropped.