fix: tokio runtime should be multi thread
There is a big in how we run background tasks such as the info fetcher and the stats exporter in the trace exporter.
We currently spawn them in the runtime held by the trace exporter. This runtime is a `current_thread`, which according to the tokio documentation only runs background task when the runtime is in a `block_on` https://docs.rs/tokio/latest/tokio/runtime/index.html#current-thread-runtime-behavior-at-the-time-of-writing.

This is called only when sending traces, which means that background tasks are paused in between traces submissions. So if the stats exporter is sending data for instance, and the runtime pauses the requests might be interrupted while it's being sent which will create issues.

The fix is to use the multi threaded runtime, which will keep a worker pool to execute tasks even when we are outside of block_on.
I configure the runtime to use only one worker thread instead of the default 4 used because we don't do have that many tasks and to be parsimonious with user resources