refactor: cleanup the telemetry configuration api (#1002)
This PR does 2 things

* Make the trace_exporter not send app-started and app-closed telemetry payloads.

Libraries already have a telemetry client sending lifecycle events and sending duplicates from the trace exporter is a bug.
There is already an API on the TelemetryWorker to drop the lifecycle events and it's juste a matter of using it. 

* Refactor the ddtelemetry builder config API

The API was very much not clear.
The spawn_with_config function took a config that would act as default, with a lower priority than fields set in the ConfigBuilder.

Using Builder::run would also try to pull configuration directly from env variables which is not good since there are other sources that could override it. This was so that integrations in the sidecar, and ddtrace dotnet would work without having to specify all configurations... but it's pretty horrible.

In practice after checking in libdatadog and other repositories I'm pretty sure this weird pattern can be replaced by putting the config in the builder and then overriding some fields as needed.

This PR also hides the `endpoint` behind a setter an a getter that properly set the telemetry path based on if direct submission is enabled or not.