fix(data-pipeline): add a wait loop in the test checking if data has been swapped (#1042)
This check is not 100% deterministic, but between the time the mock returns the response
and we swap the atomic pointer holding the agent_info we only need to perform
very few operations. We wait for a maximum of 1s before failing the test and that should
give way more time than necesssary.