[PROF-11524] Minor tweak: Set profile `start_time` after resetting (#987)
**What does this PR do?**

This PR applies a suggestion from
https://github.com/DataDog/libdatadog/pull/963#discussion_r2017375152
to set the `start_time` after obtaining the old profile.

This will hopefully make it more clear that the `start_time` is used
for the old profile, not the new one.

**Motivation:**

Make code more readable.

**Additional Notes:**

N/A

**How to test the change?**

Existing test coverage should be enough to validate this change
(for instance, the Ruby profiler test suite explicitly tests the
timestamps on profiles).