Add `reconfigure` API to crashtracker
Prior to https://github.com/DataDog/libdatadog/pull/1000, it was
possible to reconfigure the crashtracker by calling init again.

This was removed as it could cause issues with signal handlers and
chaining.

But, we still need to be able to update the configuration as for
instance the following can change at run-time for Ruby and need to
be updated:

* Endpoint
* Tags

The new `reconfigure` API accounts for this: it's basically the
same as `init` BUT without touching the signal handlers.