[PROF-11745] Allow crashtracker to be enabled/disabled + reconfigured (#1066)
# What does this PR do?

This PR adds the following new APIs to the crashtracker:
* `ddog_crasht_enable`
* `ddog_crasht_disable`
* `ddog_crasht_reconfigure`

# Motivation

Prior to https://github.com/DataDog/libdatadog/pull/1000, it was possible to:

* reconfigure the crashtracker by calling init again
* stop the crashtracker

Both features were removed as it could cause issues with signal handlers and chaining.

But, we still need to be able to update the configuration as for instance the following can change at run-time for Ruby and need to be updated:
    
* Endpoint
* Tags

This PR restores some of the behavior removed in #1000, without the unsafety of touching signal handlers:

* `enable`/`disable` just skip the crashtracker signal handling code, but never uninstall it
* `reconfigure` does the same as `init` BUT without touching the signal handlers

# Additional Notes

This PR supercedes https://github.com/DataDog/libdatadog/pull/1017 .

# How to test the change?

I've validated this branch with https://github.com/DataDog/dd-trace-rb/pull/4577 on the Ruby side, and I'm able to get a green test suite for the Ruby crashtracker.

---------

Co-authored-by: Daniel Schwartz-Narbonne <daniel.schwartznarbonne@datadoghq.com>