[PROF-11827] Enable publishing ruby gem from CI using trusted publishing (#1070)
* Fix bundler not working in rubygems/release-gem

The `rubygems/release-gem` step was failing with:

```
Run bundle exec rake release
Could not locate Gemfile or .bundle/ directory
```

...because it runs that command from the root of the repository, not the
 `ruby/` folder.

It seems a bit harder to specify the folder, so instead I'm adding a
bundler config file to tell bundler (the `bundle` command) where
to find what it needs.

* Move all files needed to release Ruby to a single folder

This way we don't need to litter the repository with random files
outside the Ruby folder.

* Enable publish step, now that the rest of the workflow is working fine

* Add logging to push step

* Simplify Ruby publish steps even further

This also fixes ruby/setup-ruby caching!

* Use `eval_gemfile` instead of copying Gemfile to root

(Because the `gems.rb` uses `gemspec` and this failed because it
wouldn't find the `gemspec` when run from the root of the repo)

* Tweak test name