[PROF-11524] Package libdatadog v18.1.0 for Ruby (#1082)
**What does this PR do?**

This PR includes the changes documented in the "Releasing a new version to rubygems.org" part of the README:
https://github.com/datadog/libdatadog/tree/main/ruby#releasing-a-new-version-to-rubygemsorg

It also includes an updated to the `gem_packaging.rb` test to add the
new `INTERNED_EMPTY_STRING` public symbol added in #917 .

**Motivation:**

Enable Ruby to use libdatadog v18.1.0. This includes A LOT of changes
to profiling, crashtracking, configuration, process discovery etc
that have been pending for Ruby for too long.

**Additional Notes:**

The `INTERNED_EMPTY_STRING` new symbol seems ok to have (e.g. I don't
expect it to clash with other libraries), but just in case I will
discuss with @danielsn prefixing it with `ddog_`.

Also it's cool to see our tests working! A new symbol that's not
prefixed with the expected strings showed up, and we caught it :)

**How to test the change?**

I've tested this release locally using the changes in
https://github.com/datadog/dd-trace-rb/pull/4577 .

As a reminder, new libdatadog releases don't get automatically picked up
by dd-trace-rb, so the PR that bumps the dependency will also test
this release against all supported Ruby versions.