initial setup of clippy-annotation-reporter github action.
This is not part of the libdatadog workspace. It's a github action that
will run for CI. It reports counts of clippy allow annotations for both
changed files in the PR and a repo overall. The goal is to call
attention to excessive usage of allows, which could be a signal of
reduced code quality. This action should be moved to its own repo when
confident of its functionality.