[NO-TICKET] Rename ffi `INTERNED_EMPTY_STRING` symbol to `ddog_INTERNED_EMPTY_STRING` (#1087)
**What does this PR do?**

This PR renames the public ffi symbol `INTERNED_EMPTY_STRING`
(introduced in #917) to `ddog_INTERNED_EMPTY_STRING`.

This is technically a breaking API change but nobody is yet using
this API
([from GitHub searches](https://github.com/search?q=org%3ADataDog%20INTERNED_EMPTY_STRING&type=code))
so doing it now is much better than doing it anytime later.

**Motivation:**

We've been trying to make sure our public ffi symbols (e.g. things
that end up in `.so` and `.dll` files and whatnot) are all prefixed
with `ddog_` to avoid clashing with any other libraries in the system.

While it's not extremely likely that `INTERNED_EMPTY_STRING` will
collide, I discussed with @danielsn that it also doesn't quite seem
worth it to create an exception for this symbol.

So this keeps the consistency with our other symbols so far.

**Additional Notes:**

N/A

**How to test the change?**

The existing example can be used to validate this change is working.