feat(data-pipeline): signal agent response payload version and only send body when changed
The datadog agent can take a `datadog-rates-payload-version` header, which signals the version of the agent rates which are sent in the agent response.
This allows not refreshing the rates in the sampler of they haven't changed.

This PR adds a feature to store, send and update this value, to return the response body from the trace exporter only if the payload  version has been updated