feat(data-pipeline):  make agent info local to the trace exporter
Having the agent info in a global variable makes some if the tests fail (when run with `cargo test`) because the agent info state is shared across all tests and tests are coupled together.

The agent info states needs to be re-usable across multiple trace exporter though because of forks.
Some languages fork very often and not actually keeping this state forces a request to the agent for every new forked process, killing performance.

In order to prevent this issue, this PR adds a new API to the trace exporter `TraceExporterPersistentState` which can be grabbed from the trace exporter before its dropped and given to the builder for recreating a new exporter.