[PROF-12141] Package libdatadog v19.1.0 for Ruby (#1146)
**What does this PR do?**

This PR includes the changes documented in the "Releasing a new version to rubygems.org" part of the README:
https://github.com/datadog/libdatadog/tree/main/ruby#releasing-a-new-version-to-rubygemsorg

**Motivation:**

Enable Ruby to use libdatadog v19.1.0. In particular, this release
speeds up crashtracking.

**Additional Notes:**

N/A

**How to test the change?**

I've tested this change locally and did not need to change
anything on the Ruby side other than the version bump. Once this
version is published, I'll open the Ruby-side PR for the update.

As a reminder, new libdatadog releases don't get automatically picked up
by dd-trace-rb, so the PR that bumps the dependency will also test
this release against all supported Ruby versions.