Fix specs failing due to slightly changed behavior on ruby-master
Some of our tests use

```ruby
let(:temporary_directory) { "does/not/exist" }
```

and in https://github.com/ruby/fileutils/pull/132 the behavior of
`FileUtils.remove_dir` was tweaked to change the exception that gets
raised when doing `FileUtils.remove_dir` with a directory that doesn't
exist.

I've updated the test clean-up hook with the new exception as well.