Replace Devflow stale branch auto-closer with MergeGate rule (#1152)
We already have a github action that handles stale PRs. The Devflow one
isn't as flexible. The intention of the rule was to require branches not
be too stale before merging. Updated the rule so that a branch should be
no stale-er than 10 days. This number is arbitrary and can be changed to
something more appropriate.