Fix removal of clients without telemetry enabled (#1176)
Clearing queue_ids happens via SidecarAction::ClearQueueId. However, these don't happen exclusively when telemetry is enabled, but also e.g. for remote config.

Avoids memory leaks in that case.

(The dd-trace-php code using this was also missing to communicate the ClearQueueId in the first place. However, before this patch, passing ClearQueueId would lead to a telemetry client to be created, which we don't want either. This is what this PR solves.)

Also, Telemetry worker spawning is neither async nor faillible. Remove the Result.

Signed-off-by: Bob Weinand <bob.weinand@datadoghq.com>