[NO-TICKET] Update build instructions in README
**What does this PR do?**

This PR does a quick pass on the "Builder crate" section of the
README.md, updating it with the latest recommended way of building
libdatadog.

**Motivation:**

I just tried to build libdatadog without the old
`build-profiling-ffi.sh` and got confused about what the steps were >_>.

**Additional Notes:**

N/A

**How to test the change?**

Run the command, observe you get a nice fresh build!