feat(trace-utils): Remove clone bound on SpanText
# Motivation

Cloning is very cheap as long as we have a ByteString or a str slice, but if we want to use python strings we need to hold the GIL to clone them.
This is not possible in general... But there are no operations requiring us to actual clone the data.

# Changes

* use an indexmap for the shared dictionary. This datastructure is exactly what we want (O(1) access and insertion order iteration) and keep only one of each value
*  Use indices into the span slice while computing top level rather than copying the service string