Reuse connections (#1249)
Refactor trace-utils to pass an http client so the connections are reused.

Refactor data-pipeline to hold a client pool so the connection with the agent is reused.

Refactor trace_flusher so it stores an http client.

Remove Option passing client to the Stats Exporter.

Merge branch 'main' into julio/reuse-connections

Co-authored-by: julio.gonzalez <julio.gonzalez@datadoghq.com>