feat(tinybytes): fix miri failures due to ptr aliasing (#1281)
# Motivation

While working on seemingly unrelated changes, I triggered a fail in MIRI tests...
I reduced the failure to the `test_from_underlying` test I added in this PR, and which fails without the other changes.

If I understand correctly, MIRI complains because the `Bytes` instance holds a static ref to the underlying pointer, while we call `drop` on it.
the stacked borrow model mentions that casting any pointer to a mutable ref, while an immutable borrow is alive in undefined behaviour.
So I think the issue is that even though we don't use the static ref after the drop has been called this still fails.

To change this, I replaced the static ref, with a pointer + length which doesn't count as a borrow.