[PROF-12941] Package libdatadog v23.0.0 for Ruby
**What does this PR do?**

This PR includes the changes documented in the "Releasing a new version
to rubygems.org" part of the README:
https://github.com/datadog/libdatadog/tree/main/ruby#releasing-a-new-version-to-rubygemsorg

**Motivation:**

Enable Ruby to use libdatadog v23.0.0. In particular, this release
switches the profile serializer to compress data using zstd instead of
lz4.

**Additional Notes:**

N/A

**How to test the change?**

I've tested this change locally with the changes in
https://github.com/DataDog/dd-trace-rb/pull/5020 and was able to get
a green test suite run for Ruby.

As a reminder, new libdatadog releases don't get automatically picked up
by dd-trace-rb, so the PR that bumps the dependency will also test this
release against all supported Ruby versions.