feat(span): remove clone bound on spantext (#1266)
feat(trace-utils): Remove clone bound on SpanText

# Motivation

Cloning is very cheap as long as we have a ByteString or a str slice, but if we want to use python strings we need to hold the GIL to clone them.
This is not possible in general... But there are no operations requiring us to actual clone the data.

# Changes

* use an indexmap for the shared dictionary. This datastructure is exactly what we want (O(1) access and insertion order iteration) and keep only one of each value
*  Use indices into the span slice while computing top level rather than copying the service string

fix: license 3rd party

feat(trace-utils): widen trace id to 128 bits

# Motivation

When we serialize we store the uppper 64 bits  of the trace id in a meta tag. But at runtime, we don't want to split it as the trace id still needs to be readable.

fix: skip high bits in 128bit trace id serialization

Merge branch 'main' into paullgdc/trace-utils/remove_clone_on_spantext

Merge branch 'main' into paullgdc/trace-utils/remove_clone_on_spantext

fix: 3rpaty + lints

Co-authored-by: paul.legranddescloizeaux <paul.legranddescloizeaux@datadoghq.com>