[NO-TICKET] Make ruby-guild codeowners of the ruby bits (#1357)
[NO-TICKET] Make ruby-guild codeowners of the ruby bits

**What does this PR do?**

This PR replaces the
[libdatadog-ruby](https://github.com/orgs/DataDog/teams/libdatadog-ruby)
_github team_ with the
[ruby-guild](https://github.com/orgs/DataDog/teams/ruby-guild) as the
codeowners of the libdatadog Ruby bits.

**Motivation:**

The libdatadog ruby bits exist solely to package libdatadog for use
in dd-trace-rb, so it makes sense to me that the same group is
responsible for both, rather than having separate, finer grained
permissions.

**Additional Notes:**

As more teams adopt libdatadog for building features for dd-trace-rb,
this configuration tweak avoids having a bottleneck when we need
to do Ruby updates.

**How to test the change?**

N/A

Co-authored-by: ivo.anjo <ivo.anjo@datadoghq.com>