tests(telemetry): flush metrics with heartbeats if the interval is small
# Motivation

Tests usually set the telemetry hearbeat interval to a small value to not have to wait to get data.
For these tests also want to get telemetry metrics fast.
This PR changes the metics flush interval to the heartbeat value if the hearbeat interval is set to a shorter duration than the default metrics flush interval.